Update
Created a autoclicker script using python. Script will eventually have a simple GUI, settings menu, and be a runnable executable that will not be triggered by an OS firewall.

Discord bot on server_data_manager.py has been updated to properly store discord server members and creating a JSON file with a list of users, added channels for the bot to pay attention in, and point system.